(fix): Move v2 modal to app level, remove timer (#420)
* chore(v2-modal): Move modal to App level to prevent it displaying multiple times within site

* chore(v2-modal): Remove modal timer

* chore(v2-modal): Cleanup
